download_videos/requirements.txt
mysql-connector-python==8.0.26
numpy==1.21.2
pandas==1.3.3
protobuf==3.18.0
python-dateutil==2.8.2
pytz==2021.1
PyYAML==5.4.1
six==1.16.0
